+++ History.md
@@ -1,3 +1,8 @@
+#### 0.2
+Users can sign up and login
+Users can add recrs
+Users can grade recs
+
 #### 0.1
 login screen with two test users
 welcome page shows name